+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## 0.9.2 - Unreleased
+## 0.10.0 - Unreleased
 
 ### Added